pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shopify-product-migration"
version = "0.1.0"
description = "Tool for migrating products between Shopify stores"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src/shopify_migration"]

[tool.pytest.ini_options]
addopts = "-vv --color=yes"
log_cli = true
log_cli_level = "INFO"
pythonpath = ["src"]
asyncio_mode = "strict"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"